bank + lessons page with all filters and pagination

+++ bank-v4/routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\AccountController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,17 +15,21 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [AccountController::class, 'index'])->name('accounts-index');
-Route::get('/create', [AccountController::class, 'create'])->name('accounts-create');
-Route::post('/', [AccountController::class, 'store'])->name('accounts-store');
+Route::get('/', [AccountController::class, 'index'])->name('accounts-index')->middleware('checkUR:user');
+Route::get('/create', [AccountController::class, 'create'])->name('accounts-create')->middleware('checkUR:admin');
+Route::post('/', [AccountController::class, 'store'])->name('accounts-store')->middleware('checkUR:admin');
 // Route::get('/account/add/{account}', [AccountController::class, 'edit'])->name('accounts-edit-add');
 // Route::get('/account/withdraw/{account}', [AccountController::class, 'edit'])->name('accounts-edit-withdraw');
 // Route::put('/account/add/{account}', [AccountController::class, 'update'])->name('accounts-update-add');
 // Route::put('/account/withdraw/{account}', [AccountController::class, 'update'])->name('accounts-update-withdraw');
 
-Route::get('/account/{addOrWithdraw}/{account}', [AccountController::class, 'edit'])->name('accounts-edit');
-Route::put('/account/{addOrWithdraw}/{account}', [AccountController::class, 'update'])->name('accounts-update');
+Route::get('/account/{addOrWithdraw}/{account}', [AccountController::class, 'edit'])->name('accounts-edit')->middleware('checkUR:admin');
+Route::put('/account/{addOrWithdraw}/{account}', [AccountController::class, 'update'])->name('accounts-update')->middleware('checkUR:admin');
 
-Route::delete('/{account}', [AccountController::class, 'destroy'])->name('accounts-delete');
+Route::delete('/{account}', [AccountController::class, 'destroy'])->name('accounts-delete')->middleware('checkUR:admin');
 
-Route::get('/show/{accountId}', [AccountController::class, 'show'])->name('account-show');
+Route::get('/show/{accountId}', [AccountController::class, 'show'])->name('account-show')->middleware('checkUR:user');
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
